+++ vitest.setup.ts
@@ -1,3 +1,4 @@
+import "@testing-library/jest-dom/vitest";
 import "jest-axe";
 import { expect } from "vitest";
 import { toHaveNoViolations } from "jest-axe";